[EasyCodingStandard] providing files moved from SourceFinder to DefaultFilesProvider
